refactor: [step3] 테스트 코드 수정 (#STEP3)
