+++ README.md
@@ -22,3 +22,4 @@
 - [X] ...... Car클래스 RacingCars로 맵핑(일급컬렉션)
 - [X] ...... 매직 넘버 수정
 - [X] ...... 인스턴스 매번 호출 이슈 수정
+- [X] ...... 테스트 코드 수정
